requirements.txt
amqp==1.4.6
BeautifulSoup==3.2.1
beautifulsoup4==4.3.2
billiard==3.3.0.20
bleach==1.4.1
boto==2.36.0
businesstime==0.2.0
celery==3.1.18
celery-haystack==0.9
cffi==0.8.2
characteristic==0.1.0
cryptography==0.9.3
dealer==2.0.4
diff-match-patch==20110725.1
dj-database-url==0.3.0
dj-static==0.0.5
Django==1.8.3
django-allauth==0.16.1
django-appconf==1.0.1
django-braces==1.4.0
django-celery==3.1.17
django-celery-transactions==0.3.2
django-classy-tags==0.5.1
django-cors-headers==0.13
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.2
django-easy-pdf==0.1.0
django-email-extras==0.3
django-filer==0.9.11
django-filter==0.11.0
django-forms-builder==0.12.2
django-grappelli==2.6.3
django-guardian==1.3
django-haystack==2.4.0
django-import-export==0.2.7
django-localflavor==1.0
django-messages==0.5.1
django-model-utils==2.2
django-modeltranslation==0.7.3
django-mptt==0.7.4
django-oauth-plus==2.2.8
django-oauth2-provider==0.2.6.1
django-paging==0.2.5
django-picklefield==0.3.1
django-polymorphic==0.7.1
django-recurrence==1.2.0
django-redis-cache==1.6.0
django-reversion==1.9.3
django-rosetta==0.7.4
django-s3-folder-storage==0.2
django-scheduler==0.7.1
django-sekizai==0.8.2
django-simple-menu==1.1.0
django-social-auth==0.7.28
django-static-compiler==0.3.3
django-storages==1.1.8
django-taggit==0.16.2
django-templatetag-sugar==0.1
django-toolbelt==0.0.1
djangorestframework==3.1.3
djangorestframework-csv==1.3.3
djangorestframework-oauth==1.0.1
djangorestframework-recursive==0.1.1
djangotoolbox==1.6.2
djrill==0.8.0
docopt==0.4.0
drf-ujson==1.1
easy-thumbnails==2.2
ecdsa==0.13
elasticsearch==1.4.0
email-reply-parser==0.3.0
Fabric==1.10.1
flower==0.7.3
future==0.9.0
geopy==1.1.3
gevent==1.0.1
gnureadline==6.3.3
greenlet==0.4.5
gunicorn==18.0
html2text==2015.6.21
html5lib==0.99999
httpagentparser==1.2.2
httpie==0.9.1
httplib2==0.8
ipython==2.0.0
jsonfield==1.0.3
kombu==3.0.26
logan==0.5.10
lxml==3.2.5
mandrill==1.0.53
nydus==0.10.7
oauth2==1.5.211
oauthlib==0.6.1
paramiko==1.15.2
pdfkit==0.4.1
phonenumbers==6.2.0
Pillow==2.9.0
pip-autoremove==0.9.0
pipdeptree==0.4.2
polib==1.0.4
psycopg2==2.5.2
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
pycrypto==2.6.1
Pygments==1.6
pynliner==0.5.1
pyOpenSSL==0.14
pyPdf==1.13
pystache==0.5.4
python-dateutil==2.2
python-gnupg==0.3.6
python-magic==0.4.6
python-memcached==1.53
python-openid==2.2.5
pytz==2015.4
queuelib==1.1.1
raven==5.0.0
redis==2.10.3
reportlab==2.7
requests==2.3.0
requests-oauthlib==0.4.0
s3cmd==1.5.2
Scrapy==0.22.2
service-identity==1.0.0
setproctitle==1.1.8
shortuuid==0.4.2
simple-crypt==3.0.2
simplejson==3.3.2
six==1.5.2
sphinx-me==0.3
sqlparse==0.1.11
static==1.0.2
tablib==0.9.11
twilio==3.8.0
ujson==1.33
Unidecode==0.4.18
urllib3==1.11
vobject==0.8.1rc0
w3lib==1.6
xhtml2pdf==0.0.6
xlrd==0.9.3
xmltodict==0.9.0
